Update UI CMS install
Add simple Render.php, Router.php, Checker.php, Controller.php classes.
Create simple MVC system.
