+++ install/locale/en/install.php
@@ -55,14 +55,14 @@ return [
     'errors'                   => 'Errors',
     'final'                    => 'Finish',
     'final_note'               => 'Do not forget to delete the folder <span class="red"><b>/install</b></span>',
-    'final_warning'            => '<h2 class="red">WARNING!</h2><h3>After installation, you must delete folder /<span class="red">install</span></h3>',
+    'final_warning'            => 'WARNING!<br>After installation, you must delete folder/<span style="color:red;">install</span>',
     'go_to_site'               => 'Go to Site',
     'home'                     => 'Home',
     'ignore_warnings'          => 'Ignore all the warnings and start the installation',
     'install'                  => 'Installation',
     'install_demo'             => 'Install the DEMO data (on Russian)',
     'install_demo_help'        => 'Will be installed the basic setup and some texts (in Russian)',
-    'install_note'             => '<h3 class="green">Welcome!</h3><ul><li>List of changes located in the file <a href="?act=changelog">CHANGELOG.md</a></li><li>Additional information is available on the official project website <a href="http://johncms.com">http://johncms.com</a></li><li>Installing and using JohnCMS, means complete acceptance of the terms of the <a href="?act=license">license</a></li></ul>',
+    'install_note'             => '<p><strong class="d-block text-center">Welcome!</strong><br>List of changes located in the file <a href="?act=changelog">CHANGELOG.md</a><br>Additional information is available on the official project website <a href="http://johncms.com">http://johncms.com</a><br>Installing and using JohnCMS, means complete acceptance of the terms of the <a href="?act=license">license</a><br></p>',
     'installation_completed'   => 'Installation successful<br />Now you can go to the admin panel and customize your site',
     'languages'                => 'After installation, you can add additional languages.<br />They can be downloaded from <a href="http://johncms.com/download/index.php?cat=483">JohnCMS</a>',
     'not_recommended'          => '(not recommended)',
@@ -90,4 +90,6 @@ return [
     'update_warning'           => '<h2 class="red">ATTENTION!</h2><ul><li>Please note that an update is possible only for the original (no mods) system <b>JohnCMS #UPDATE_VERSION#</b><br />If you use any Mods, the possibility of updating necessarily agree with their authors.<br />Installing this update on the modified system can lead to complete Site failure.</li><li>Before starting the upgrade process, save a backup copy of the database.<br />If for some reason the update is not finished, you have to restore the database from backup.</li><li>Some steps can take a long time (several minutes), which depends on the size of the database,  hosting, etc.</li><li>If you begin your upgrade, then undo the changes will not be possible without restoring from a backup</li><li>Be sure to read the <a href="../#MODE#.txt">instructions for updating</a>.</li></ul>',
     'updated'                  => 'Updated',
     'warnings'                 => 'Warnings',
+    'step'                     => 'Step',
+    'powered_by'               => 'Powered by',
 ];
