+++ install/locale/ru/install.php
@@ -62,7 +62,7 @@ return [
     'install'                  => 'Установка',
     'install_demo'             => 'Установить ДЕМО данные',
     'install_demo_help'        => 'Будут установлены базовые настройки и некоторые тексты (на Русском)',
-    'install_note'             => '<h3 class="green">Добро пожаловать!</h3><ul><li>Список изменений находится в файле <a href="?act=changelog">CHANGELOG.md</a></li><li>Дополнительную информацию Вы можете получить на официальном сайте <a href="http://johncms.com">http://johncms.com</a></li><li>Установка и использование JohnCMS, означает полное согласие с условиями <a href="?act=license">лицензии</a></li></ul>',
+    'install_note'             => '<p> <strong class="d-block text-center">Добро пожаловать!</strong> <br>Список изменений находится в файле <a href="?act=changelog">CHANGELOG.md</a> <br>Дополнительную информацию Вы можете получить на официальном сайте <a href="http://johncms.com">http://johncms.com</a> <br>Установка и использование JohnCMS, означает полное согласие с условиями <a href="?act=license">license</a> <br></p>',
     'installation_completed'   => 'Установка успешно завершена<br />Теперь можно зайти в Админ панель и настроить свой Сайт',
     'languages'                => 'Дополнительные языки вы сможете добавить после установки.<br />Скачать их можно с сайта <a href="http://johncms.com/download/index.php?cat=483">JohnCMS</a>',
     'not_recommended'          => '(не рекомендуется)',
@@ -90,4 +90,6 @@ return [
     'update_warning'           => '<h2 class="red">ВНИМАНИЕ!</h2><ul><li>Учтите, что обновление возможно только для оригинальной (без модов) системы <b>JohnCMS #UPDATE_VERSION#</b><br />Если Вы используете какие-либо моды, то возможность обновления обязательно согласуйте с их авторами.<br />Установка данного обновления на модифицированную систему может привести к полной неработоспособности сайта.</li><li>Перед началом процедуры обновления ОБЯЗАТЕЛЬНО сделайте резервную копию базы данных.<br />Если по какой-то причине обновление не пройдет до конца, Вам придется восстанавливать базу из резервной копии.</li><li>Некоторые этапы обновления могут занимать довольно продолжительное время (несколько минут), которое зависит от размера базы данных сайта и скорости сервера хостинга.</li><li>Если Вы начнете процедуру обновления, то отмена изменений будет невозможна без восстановления из резервной копии</li><li>Обязательно ознакомьтесь с <a href="../#MODE#.txt">инструкцией по обновлению</a>.</li></ul>',
     'updated'                  => 'Обновлен',
     'warnings'                 => 'Предупреждения',
+    'step'                     => 'Шаг',
+    'powered_by'               => 'Работает на',
 ];
