+++ install/src/update_800_900.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 const DEBUG = true;
 const _IN_JOHNCMS = true;
 
-require '../system/bootstrap.php';
+require __DIR__ . '/../../system/bootstrap.php';
 
 /** @var PDO $db */
 $db = di(PDO::class);
