+++ install/views/layouts/base.html
@@ -1,0 +1,116 @@
+<!doctype html>
+<html lang="{{ lang }}">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Install</title>
+
+    <style>
+        body {
+            font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol";
+            background: #f9f9f9;
+            padding: 0;
+            margin: 0;
+            color: #313131;
+            box-sizing: border-box;
+        }
+
+        header {
+            box-shadow: 0 1px 0 1px rgba(0, 0, 0, .1);
+            padding: 16px;
+            background: #fff;
+            font-weight: bold;
+            font-size: 24px;
+            margin-bottom: 10px;
+        }
+
+        main {
+            display: flex;
+            flex-direction: column;
+            align-items: center;
+            justify-content: center;
+            min-height: calc(100vh - 125px);
+        }
+
+        footer {
+            padding: 15px;
+            background: #1d2124;
+            color: #fff;
+        }
+
+        footer a {
+            color: #fff1b8;
+        }
+
+        p {
+            margin-top: 15px;
+        }
+
+        label {
+            display: block;
+            font-size: small;
+            font-weight: bold;
+            margin-top: 10px;
+        }
+
+        .text-muted{
+            color:#6c757d
+        }
+
+        .card {
+            background: #fff;
+            border-radius: 3px;
+            max-width: 90%;
+            box-shadow: 0 1px 5px 0 rgba(0, 0, 0, .1);
+        }
+
+        .card-header {
+            background: #fff;
+            border-bottom: solid 1px #f6f6f6;
+            padding: 10px;
+            font-weight: bold;
+        }
+
+        .card-content {
+            padding: 10px;
+        }
+
+        .card-actions {
+            text-align: right;
+            padding: 10px;
+            border-top: solid 1px #f6f6f6;
+        }
+
+        .btn {
+            background: #fff;
+            border: solid 1px #efefef;
+            cursor: pointer;
+            padding: 6px 12px;
+        }
+
+        .btn:focus, .btn:hover {
+            box-shadow: 0 0 1px 1px rgba(0, 7, 255, 0.3);
+        }
+
+        .text-center {
+            text-align: center;
+        }
+
+        .d-block {
+            display: block;
+        }
+
+        .text-green {
+            color: #2c6a57
+        }
+    </style>
+</head>
+<body>
+    <header>{{ lang_arr['install'] }} JohnCMS</header>
+    <main>
+        {{ content }}
+    </main>
+    <footer>&copy;&#160;{{ lang_arr['powered_by'] }} <a href="https://johncms.com">JohnCMS</a></footer>
+</body>
+</html>
