+++ install/views/templates/error.html
@@ -1,0 +1,12 @@
+<style>
+    .card-header {
+        background: #ff1b2f;
+        border-bottom: solid 1px #791522;
+        color: white;
+    }
+</style>
+
+<div class="card">
+    <div class="card-header">Error: {{ error_title }}</div>
+    <div class="card-content">{{ error_message }}</div>
+</div>
