+++ install/views/templates/text.html
@@ -1,0 +1,1 @@
+<div style="padding: 10px">{{ text }}</div>
